Merging with main

+++ solar_realtime_pipeline.py
@@ -3,6 +3,7 @@ import os, sys, glob, getopt
 os.environ["MKL_NUM_THREADS"] = "1"
 os.environ["NUMEXPR_NUM_THREADS"] = "1"
 os.environ["OMP_NUM_THREADS"] = "1"
+os.environ["OPENBLAS_NUM_THREADS"] = "1"
 from signal import signal, SIGPIPE, SIG_DFL
 signal(SIGPIPE,SIG_DFL)
 from ovrolwasolar import solar_pipeline as sp
@@ -93,8 +94,10 @@ def sun_riseset(date=Time.now(), observatory='ovro', altitude_limit=15.):
     except Exception as e:
         logging.error(e)
 
+    t_now = Time(date_mjd, format='mjd')
+
     obs = EarthLocation.of_site(observatory)
-    t0 = Time(int(date_mjd) + 13. / 24., format='mjd')
+    t0 = Time(int(date_mjd) + 8. / 24., format='mjd')
     
     sun_loc = get_body('sun', t0, location=obs)
     
@@ -103,18 +106,19 @@ def sun_riseset(date=Time.now(), observatory='ovro', altitude_limit=15.):
         t0 += TimeDelta(60., format='sec')
         alt = sun_loc.transform_to(AltAz(obstime=t0, location=obs)).alt.degree
 
-    t1 = Time(int(date_mjd) + 22. / 24., format='mjd')
+    t1 = Time(int(date_mjd) + 20. / 24., format='mjd')
     sun_loc = get_body('sun', t1, location=obs)
     alt = sun_loc.transform_to(AltAz(obstime=t1, location=obs)).alt.degree
     while alt > altitude_limit:
         t1 += TimeDelta(60., format='sec')
         alt = sun_loc.transform_to(AltAz(obstime=t1, location=obs)).alt.degree
 
+    if t_now < t1-TimeDelta(1., format='jd'):
+        # still before previous day sunset
+        t1 = t1 - TimeDelta(1., format='jd')
+        t0 = t0 - TimeDelta(1., format='jd')
+
     return t0, t1
-
-
-
-
 
 
         
@@ -161,10 +165,10 @@ def check_fast_ms(msname):
         return False
     return True
 
-    
-def run_calib(msfile, msfiles_cal=None, bcal_tables=None, do_selfcal=True, num_phase_cal=0, 
+
+def run_calib(msfile, msfiles_cal=None, bcal_tables=None, do_selfcal=True, num_phase_cal=0, slowfast='slow',
                 num_apcal=1, caltable_folder=None, logger_file=None, visdir_slfcaled=None, 
-                flagdir=None, delete_allsky=False, actively_rm_ms=True, stokes='I'):
+                flagdir=None, delete_allsky=False, actively_rm_ms=True, stokes='I', manual_flagging_ants=None):
     
     try:
         msmd.open(msfile)
@@ -173,17 +177,31 @@ def run_calib(msfile, msfiles_cal=None, bcal_tables=None, do_selfcal=True, num_p
         return -1
 
     # do time average if the input ms file is fast visibility
-    if check_fast_ms(msfile):
-        omsfile = os.path.dirname(msfile) + '/' + os.path.basename(msfile).replace('.ms', '.10s.ms')
-        split(msfile, omsfile, datacolumn='data', timebin='10s')
-        os.system('rm -rf ' + msfile)
-        os.system('mv ' + omsfile + ' ' + msfile)
+    if slowfast == 'fast':
+        if check_fast_ms(msfile):
+            omsfile = os.path.dirname(msfile) + '/' + os.path.basename(msfile).replace('.ms', '.10s.ms')
+            split(msfile, omsfile, datacolumn='data', timebin='10s')
+            os.system('rm -rf ' + msfile)
+            os.system('mv ' + omsfile + ' ' + msfile)
 
-    msmd.open(msfile)
-    trange = msmd.timerangeforobs(0)
-    btime = qa.time(trange['begin']['m0'],form='fits')[0]
-    etime = qa.time(trange['end']['m0'],form='fits')[0]
-    msmd.close()
+    if manual_flagging_ants is not None:
+        # Flag the specified antennas
+        try:
+            flagdata(vis=msfile, mode='manual', antenna=manual_flagging_ants, action='apply')
+            logging.info(f"Manually flagged antennas: {manual_flagging_ants}")
+        except Exception as e:
+            logging.error(f"Error flagging antennas {manual_flagging_ants}: {e}")
+
+    
+    # var not used
+
+    #msmd.open(msfile)
+    #trange = msmd.timerangeforobs(0)
+    #btime = qa.time(trange['begin']['m0'],form='fits')[0]
+    #etime = qa.time(trange['end']['m0'],form='fits')[0]
+    #msmd.close()
+
+
     cfreqidx = os.path.basename(msfile).find('MHz') - 2
     cfreq = os.path.basename(msfile)[cfreqidx:cfreqidx+2]+'MHz'
     msfile_cal_ = [m for m in msfiles_cal if cfreq in m]
@@ -220,6 +238,7 @@ def run_calib(msfile, msfiles_cal=None, bcal_tables=None, do_selfcal=True, num_p
                                         fast_vis=fast_vis, delete_allsky=delete_allsky,\
                                         fast_vis_image_model_subtraction=fast_vis_image_model_subtraction,\
                                         sky_image=sky_image,pol=stokes)
+            
             if actively_rm_ms:
                 os.system('rm -rf ' + msfile)
             os.system('cp -r '+ outms + ' ' + visdir_slfcaled + '/')
@@ -242,7 +261,7 @@ def run_calib(msfile, msfiles_cal=None, bcal_tables=None, do_selfcal=True, num_p
             outms, tmp = sp.image_ms_quick(msfile, calib_ms=msfile_cal, do_selfcal=do_selfcal, imagename=imagename, logging_level='info', 
                         num_phase_cal=num_phase_cal, num_apcal=num_apcal,
                         logfile=logger_file, caltable_folder=caltable_folder, do_final_imaging=False, 
-                        do_fluxscaling=False, freqbin=1, delete_allsky=delete_allsky)
+                        do_fluxscaling=False, freqbin=1, delete_allsky=delete_allsky, use_selfcal_img_to_subtract=True)
             os.system('cp -r '+ outms + ' ' + visdir_slfcaled + '/')
             msfile_slfcaled = visdir_slfcaled + '/' + os.path.basename(outms)
             return msfile_slfcaled
@@ -673,18 +692,14 @@ def daily_beam_correction(date, save_dir='/lustre/solarpipe/realtime_pipeline/',
                 os.makedirs(hdf_dir1 + datedir)
     
             for str1 in ['fch','mfs']:
-                if os.path.exists(dir1 + datedir0):
-                    fits_lv_d0 = glob.glob(dir1 + datedir0 + '*'+str1+'*.fits')
+                if os.path.exists(dir1 + datedir):
+                    fits_lv_d0 = glob.glob(dir1 + datedir + '*'+str1+'*.fits')
                     fits_lv_d0.sort()
                 else:
                     fits_lv_d0=[]
-                if os.path.exists(dir1 + datedir1):
-                    fits_lv_d1 = glob.glob(dir1 + datedir1 + '*'+str1+'*.fits')
-                    fits_lv_d1.sort()
-                else:
-                    fits_lv_d1=[]
+                
                 # narrow down to all files between 13 UT and 03 UT of the second day 
-                fits_lv_all_ = fits_lv_d0 + fits_lv_d1
+                fits_lv_all_ = fits_lv_d0
                 fits_lv_all = []
 
         
@@ -698,12 +713,19 @@ def daily_beam_correction(date, save_dir='/lustre/solarpipe/realtime_pipeline/',
                         fits_lv_all.append(f)
                 
                 for img in fits_lv_all:
+                
+                    with fits.open(img) as hdu:
+                    
+                        datetimestr=hdu[0].header['DATE-OBS']
+                        datedir_path=(Time(datetimestr).datetime).strftime("%Y/%m/%d/")
+                        
+                    hdf_file=os.path.join(hdf_dir1,datedir_path,os.path.basename(img).replace('.fits','.hdf'))
+                    if os.path.isfile(hdf_file):
+                        continue
                     try:
                         correct_primary_beam_self_terms(img,pol=stokes)
-                        with fits.open(img) as hdu:
-                            datetimestr=hdu[0].header['DATE-OBS']
-                            datedir_path=(Time(datetimestr).datetime).strftime("%Y/%m/%d/")
-                        hdf_file=os.path.join(hdf_dir1,datedir_path,os.path.basename(img).replace('.fits','.hdf'))
+                        
+                        
                         print (img,hdf_file)
                         utils.compress_fits_to_h5(img, hdf_file)
                         print ("--------------hdf5 file written-------------------")
@@ -810,11 +832,11 @@ def correct_primary_beam_self_terms(imagename, pol='I'):
         
 
 def pipeline_quick(image_time=Time.now() - TimeDelta(20., format='sec'), server=None, lustre=True, file_path='slow', 
-            min_nband=6, nch_out=12, beam_fit_size=2, briggs=-0.5, stokes='I', do_selfcal=True, num_phase_cal=0, num_apcal=1, 
+            min_nband=6, nch_out=12, beam_fit_size=2, briggs=-0.5, stokes='I', do_selfcal=True, num_phase_cal=1, num_apcal=0, 
             overwrite_ms=False, delete_ms_slfcaled=False, logger_file=None, compress_fits=True,
-            proc_dir_mem = '/dev/shm/srtmp/', proc_dir = '/fast/solarpipe/realtime_pipeline/',
+            proc_dir_mem = '/dev/shm/srtmp/', proc_dir ='/dev/shm/srtmp/',# '/fast/solarpipe/realtime_pipeline/',
             save_dir = '/lustre/solarpipe/realtime_pipeline/',
-            calib_dir = '/lustre/solarpipe/realtime_pipeline/caltables/',
+            calib_dir = '/fast/solarpipe/caltables/',
             calib_file = '20240117_145752',
             delete_working_ms=True, delete_working_fits=True, do_refra=True, overbright=2e6, save_selfcaltab=False,
             slowfast='slow', do_imaging=True, delete_allsky=False, save_allsky=False,
@@ -897,6 +919,8 @@ def pipeline_quick(image_time=Time.now() - TimeDelta(20., format='sec'), server=
 
     bcal_tables = glob.glob(caltable_folder + calib_file + '_*MHz.bcal')
     bcal_tables.sort()
+
+    logging.debug('Found {0:d} bandpass calibration tables with {1:s}'.format(len(bcal_tables), calib_file))
 
     if not os.path.exists(visdir_work):
         os.makedirs(visdir_work)
@@ -1048,6 +1072,8 @@ def pipeline_quick(image_time=Time.now() - TimeDelta(20., format='sec'), server=
             
             # parallelized calibration, selfcalibration, and source subtraction
             logging.debug('Starting to calibrate all {0:d} bands'.format(len(msfiles)))
+            logging.debug('Num of bcal {0:d} bands'.format(len(bcal_tables)))
+
             time_cal1 = timeit.default_timer()
 
             #result = pool.map_async(run_calib, msfiles)
@@ -1570,14 +1596,14 @@ def remove_old_items(directory=".", minutes=45):
                 pass
         
 
-def run_pipeline(time_start=Time.now(), time_end=None, time_interval=600., delay_from_now=180., do_selfcal=True, num_phase_cal=0, num_apcal=1, 
+def run_pipeline(time_start=Time.now(), time_end=None, time_interval=600., delay_from_now=180., do_selfcal=True, num_phase_cal=1, num_apcal=0, 
         server=None, lustre=True, file_path='slow', multinode=True, slurmmanaged=True, taskids='0123456789', delete_ms_slfcaled=True, slowfast='slow', 
         logger_dir = '/lustre/solarpipe/realtime_pipeline/logs/', logger_prefix='solar_realtime_pipeline', logger_level=20,
         #proc_dir = '/fast/solarpipe/realtime_pipeline/',
         proc_dir_mem = '/dev/shm/srtmp/', proc_dir = '/fast/solarpipe/realtime_pipeline/',
         proc_dir_isolation = True,
         save_dir = '/lustre/solarpipe/realtime_pipeline/',
-        calib_dir = '/lustre/solarpipe/realtime_pipeline/caltables/',
+        calib_dir = '/fast/solarpipe/caltables/',
         calib_file = '20240117_145752', altitude_limit=10., 
         beam_fit_size = 2,
         briggs=-0.5,
@@ -1785,7 +1811,7 @@ def run_pipeline(time_start=Time.now(), time_end=None, time_interval=600., delay
                 print ("Doing leakage correction")
                 daily_leakage_correction(date_synop,save_dir=save_dir,overwrite=False, leakage_database=leakage_database,stokes=stokes)
                 
-            daily_beam_correction(date_synop, save_dir=savedir, overwrite=False,stokes=stokes)
+            daily_beam_correction(date_synop, save_dir=save_dir, overwrite=False,stokes=stokes)
                 
             if slowfast.lower() == 'fast':
                 twait += TimeDelta(600., format='sec') 
@@ -1836,6 +1862,7 @@ if __name__=='__main__':
         pdsh -w lwacalim[00-09] 'rm -rf /dev/shm/srtmp/*' # clear ram disk
         pdsh -w lwacalim[00-09] 'pkill -u solarpipe -f wsclean'
         pdsh -w lwacalim[00-09] 'pkill -u solarpipe -f python'
+        pdsh -w lwacalim[00-09] 'find /dev/shm -maxdepth 1 -name "__KMP_REGISTERED_LIB*" -exec rm -f {} + 2>/dev/null'
     """
     parser = argparse.ArgumentParser(description='Solar realtime pipeline')
     parser.add_argument('--start_time', default=Time.now().isot, type=str, help='Timestamp for the start time. Format YYYY-MM-DDTHH:MM')
@@ -1847,9 +1874,10 @@ if __name__=='__main__':
     parser.add_argument('--nolustre', default=False, help='If set, do NOT assume that the data are stored under /lustre/pipeline/ in the default tree', action='store_true')
     parser.add_argument('--file_path', default='slow/', help='Specify where the raw data is located')
     parser.add_argument('--proc_dir_mem', default='/dev/shm/srtmp/' )# default='/fast/solarpipe/realtime_pipeline/', help='Directory for processing')
+#    parser.add_argument('--proc_dir_mem', default='/fast/solarpipe/realtime_pipeline/proc/', help='Directory for processing')
     parser.add_argument('--proc_dir', default='/fast/solarpipe/realtime_pipeline/', help='Directory for processing')
     parser.add_argument('--save_dir', default='/lustre/solarpipe/realtime_pipeline/', help='Directory for saving fits files')
-    parser.add_argument('--calib_dir', default='/lustre/solarpipe/realtime_pipeline/caltables/', help='Directory to calibration tables')
+    parser.add_argument('--calib_dir', default='/fast/solarpipe/caltables/', help='Directory to calibration tables')
     parser.add_argument('--calib_file', default='', help='Calibration file to be used yyyymmdd_hhmmss')
     parser.add_argument('--alt_limit', default=15., help='Lowest solar altitude to start/end imaging')
     parser.add_argument('--bmfit_sz', default=2, help='Beam fitting size to be passed to wsclean')
@@ -1862,7 +1890,7 @@ if __name__=='__main__':
     parser.add_argument('--keep_working_ms', default=False, help='If True, keep the working ms files after imaging', action='store_true')
     parser.add_argument('--keep_working_fits', default=False, help='If True, keep the working fits files after imaging', action='store_true')
     parser.add_argument('--save_allsky', default=False, help='If True, save the band-averaged all sky images', action='store_true')
-    parser.add_argument('--save_selfcaltab', default=True, help='If True, save the selfcalibration tables', action='store_true')
+    parser.add_argument('--save_selfcaltab', default=False, help='If True, save the selfcalibration tables', action='store_true')
     parser.add_argument('--no_selfcal', default=False, help='If set, do not do selfcal regardless slow or fast', action='store_true')
     parser.add_argument('--no_imaging', default=False, help='If set, do not perform imaging', action='store_true')
     parser.add_argument('--nonstop', default=False, help='If set, the script will be run without stopping', action='store_true')
@@ -1895,11 +1923,14 @@ if __name__=='__main__':
         logging.info('Calibration tables not provided or recognized. Attempting to find those from default location on lwacalim.')
         calib_tables = glob.glob('/lustre/solarpipe/realtime_pipeline/caltables_latest/*.bcal')
         if len(calib_tables) > 10:
-            calib_file = os.path.basename(calib_tables[0])[:15]
+            calib_file = '_'.join(os.path.basename(calib_tables[0]).split('_')[0:-1])
             logging.info('Using calibration file {0:s}'.format(calib_file))
         else:
             logging.error('No calibration files found. Abort.')
             sys.exit(0)
+
+    logging.info('Calibration file to be used: {0:s}'.format(calib_file))
+    print('Calibration file to be used: {0:s}'.format(calib_file))
 
     try:
         run_pipeline(time_start=args.start_time, time_end=Time(args.end_time), time_interval=float(args.interval), taskids=args.taskids, delay_from_now=float(args.delay),
